idp/poc/backend/requirements.txt
Flask==3.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.27
google-cloud-secret-manager==2.18.0
firebase-admin==6.4.0 